+++ R/tdf.R
@@ -2,10 +2,10 @@
 #'
 #' This function returns `TRUE` for the object returned from the 
 #' \code{generate.placebos} function.
-#' and \code{FALSE} for all other objects, including regular data frames.
+#' and `FALSE` for all other objects, including regular data frames.
 #'
 #' @param x An object
-#' @return \code{TRUE} if the object inherits from the `tdf` class.
+#' @return `TRUE` if the object inherits from the `tdf` class.
 #' @export
 is_tdf <- function(x) {
 	inherits(x, "tdf")
@@ -15,10 +15,10 @@ is_tdf <- function(x) {
 #'
 #' This function returns `TRUE` for the object returned from the 
 #' \code{multiple.synth} function.
-#' and \code{FALSE} for all other objects, including regular data frames.
+#' and `FALSE` for all other objects, including regular data frames.
 #'
 #' @param x An object
-#' @return \code{TRUE} if the object inherits from the `tdf_multi` class.
+#' @return `TRUE` if the object inherits from the `tdf_multi` class.
 #' @export
 is_tdf_multi <- function(x) {
 	inherits(x, "tdf_multi")