+++ NEWS.md
@@ -1,3 +1,8 @@
+# SCtools 0.3.1
+* Changed `plot_placebos` to have the vertical line at the treatment time instead of first post-treatment time point
+* Fixed the p-value calculation in `plac_dist`
+* Minor fixes to speed up `multiple.synth` including enabling multiprocess
+
 # SCtools 0.3.0
 
 * Added a `NEWS.md` file to track changes to the package.